Make GridRenderer responsible for rendering grids and coordinates
`Grid#draw()` and `Grid#drawCoordinatesOverlay()` are no longer used and have been removed (along with much other unused
Grid functionality).
