+++ src/main/java/net/rptools/maptool/model/SquareGrid.java
@@ -14,17 +14,11 @@
  */
 package net.rptools.maptool.model;
 
-import java.awt.Color;
 import java.awt.Dimension;
-import java.awt.Font;
-import java.awt.FontMetrics;
-import java.awt.Graphics2D;
 import java.awt.Point;
 import java.awt.Rectangle;
 import java.awt.event.KeyEvent;
 import java.awt.geom.Area;
-import java.awt.geom.Line2D;
-import java.awt.geom.Path2D;
 import java.awt.geom.Point2D;
 import java.awt.image.BufferedImage;
 import java.util.HashMap;
@@ -32,16 +26,11 @@ import java.util.List;
 import java.util.Map;
 import javax.swing.Action;
 import javax.swing.KeyStroke;
-import javax.swing.SwingUtilities;
 import net.rptools.maptool.client.AppPreferences;
 import net.rptools.maptool.client.MapTool;
-import net.rptools.maptool.client.ScreenPoint;
-import net.rptools.maptool.client.swing.SwingUtil;
 import net.rptools.maptool.client.tool.PointerTool;
 import net.rptools.maptool.client.ui.theme.Images;
 import net.rptools.maptool.client.ui.theme.RessourceManager;
-import net.rptools.maptool.client.ui.zone.renderer.GridRenderer;
-import net.rptools.maptool.client.ui.zone.renderer.ZoneRenderer;
 import net.rptools.maptool.client.walker.WalkerMetric;
 import net.rptools.maptool.client.walker.ZoneWalker;
 import net.rptools.maptool.client.walker.astar.AStarSquareEuclideanWalker;
@@ -49,7 +38,7 @@ import net.rptools.maptool.language.I18N;
 import net.rptools.maptool.server.proto.GridDto;
 import net.rptools.maptool.server.proto.SquareGridDto;
 
-public class SquareGrid extends Grid {
+public final class SquareGrid extends Grid {
   private static final String alpha = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"; // $NON-NLS-1$
   private static final Dimension CELL_OFFSET = new Dimension(0, 0);
   private static final BufferedImage pathHighlight =
@@ -229,72 +218,6 @@ public class SquareGrid extends Grid {
   }
 
   @Override
-  public void drawCoordinatesOverlay(Graphics2D g, ZoneRenderer renderer) {
-    Object oldAA = SwingUtil.useAntiAliasing(g);
-
-    Font oldFont = g.getFont();
-    g.setFont(g.getFont().deriveFont(20f).deriveFont(Font.BOLD));
-    FontMetrics fm = g.getFontMetrics();
-
-    double cellSize = renderer.getScaledGridSize();
-    CellPoint topLeft =
-        convert(new ScreenPoint(0, 0).convertToZone(renderer.getViewModel().getZoneScale()));
-    var topLeftZone = convert(topLeft);
-    ScreenPoint sp =
-        renderer.getViewModel().getZoneScale().toScreenSpace(topLeftZone.x, topLeftZone.y);
-
-    Dimension size = renderer.getSize();
-
-    int startX = SwingUtilities.computeStringWidth(fm, "MMM") + 10;
-
-    double x = sp.x + cellSize / 2; // Start at middle of the cell that's on screen
-    int nextAvailableSpace = -1;
-    while (x < size.width) {
-      String coord = Integer.toString(topLeft.x);
-
-      int strWidth = SwingUtilities.computeStringWidth(fm, coord);
-      int strX = (int) x - strWidth / 2;
-
-      if (x > startX && strX > nextAvailableSpace) {
-        g.setColor(Color.black);
-        g.drawString(coord, strX - 1, fm.getHeight() - 1);
-        g.drawString(coord, strX + 1, fm.getHeight() - 1);
-        g.drawString(coord, strX - 1, fm.getHeight() + 1);
-        g.drawString(coord, strX + 1, fm.getHeight() + 1);
-        g.setColor(Color.orange);
-        g.drawString(coord, strX, fm.getHeight());
-
-        nextAvailableSpace = strX + strWidth + 10;
-      }
-      x += cellSize;
-      topLeft.x++;
-    }
-    double y = sp.y + cellSize / 2; // Start at middle of the cell that's on screen
-    nextAvailableSpace = -1;
-    while (y < size.height) {
-      String coord = decimalToAlphaCoord(topLeft.y);
-
-      int strY = (int) y + fm.getAscent() / 2;
-
-      if (y > fm.getHeight() && strY > nextAvailableSpace) {
-        g.setColor(Color.black);
-        g.drawString(coord, 10 - 1, strY - 1);
-        g.drawString(coord, 10 + 1, strY - 1);
-        g.drawString(coord, 10 - 1, strY + 1);
-        g.drawString(coord, 10 + 1, strY + 1);
-        g.setColor(Color.yellow);
-        g.drawString(coord, 10, strY);
-
-        nextAvailableSpace = strY + fm.getAscent() / 2 + 10;
-      }
-      y += cellSize;
-      topLeft.y++;
-    }
-    g.setFont(oldFont);
-    SwingUtil.restoreAntiAliasing(g, oldAA);
-  }
-
-  @Override
   public boolean useMetric() {
     return true;
   }
@@ -385,42 +308,6 @@ public class SquareGrid extends Grid {
   @Override
   public GridCapabilities getCapabilities() {
     return CAPABILITIES;
-  }
-
-  @Override
-  public void draw(ZoneRenderer renderer, Graphics2D g, Rectangle bounds) {
-    var zoneScale = renderer.getViewModel().getZoneScale();
-    double scale = zoneScale.getScale();
-    double gridSize = getSize() * scale;
-
-    g.setColor(new Color(getZone().getGridColor()));
-
-    int offX = (int) (zoneScale.getOffsetX() % gridSize + getOffsetX() * scale);
-    int offY = (int) (zoneScale.getOffsetY() % gridSize + getOffsetY() * scale);
-
-    int startCol = (int) ((int) (bounds.x / gridSize) * gridSize);
-    int startRow = (int) ((int) (bounds.y / gridSize) * gridSize);
-    Path2D path = new Path2D.Double();
-    for (double row = startRow; row < bounds.y + bounds.height + gridSize; row += gridSize) {
-      path.append(
-          new Line2D.Double(
-              bounds.x, (int) (row + offY), bounds.x + bounds.width, (int) (row + offY)),
-          false);
-    }
-    for (double col = startCol; col < bounds.x + bounds.width + gridSize; col += gridSize) {
-      path.append(
-          new Line2D.Double(
-              (int) (col + offX), bounds.y, (int) (col + offX), bounds.y + bounds.height),
-          false);
-    }
-    GridRenderer.drawGridShape(zoneScale, getSize(), g, path);
-  }
-
-  public ZonePoint getCenterPoint(CellPoint cellPoint) {
-    ZonePoint zp = convert(cellPoint);
-    zp.x += (int) (getCellWidth() / 2d);
-    zp.y += (int) (getCellHeight() / 2d);
-    return zp;
   }
 
   public static String decimalToAlphaCoord(int value) {
